backend-api/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-jose
pydantic
sqlalchemy
python-dotenv 